Fixed HotelTypes controller

+++ HotelManagerSystem.API/AuthBL/Controllers/DirectoriesController/HotelTypesController.cs
@@ -1,10 +1,15 @@
-﻿using HotelManagerSystem.BL.Directories;
+﻿using HotelManagerSystem.API.Responses;
+using HotelManagerSystem.BL.Directories;
+using HotelManagerSystem.BL.Exceptions;
 using HotelManagerSystem.DAL.Responses;
+using HotelManagerSystem.Models.Data;
+using HotelManagerSystem.Models.DTOs;
 using HotelManagerSystem.Models.Entities;
 using HotelManagerSystem.Models.Request.CreateRequest;
 using HotelManagerSystem.Models.Request.UpdateRequest;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 
 namespace HotelManagerSystem.API.AuthBL.Controllers.DirectoriesController
 {
@@ -24,63 +29,73 @@ namespace HotelManagerSystem.API.AuthBL.Controllers.DirectoriesController
 
         [HttpPost]
         [Route("create")]
-        public async Task<Response> Create([FromBody] CreateNameDirectoryRequest request)
+        public async Task<ActionResult> Create([FromBody] CreateNameDirectoryRequest request)
         {
             try
             {
-                await _service.Create(request);
+                var hotelType = await _service.Create(request);
+                return Created("", hotelType);
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error while processing request from {Hotel Types}", request);
+                return BadRequest(new ErrorResponse(ex.Message));
             }
-
-            return new Response(200, true, null);
         }
 
         [HttpPut]
         [Route("update")]
-        public async Task<Response> Update(UpdateNameDirectoryRequest request)
+        public async Task<ActionResult> Update(UpdateNameDirectoryRequest request)
         {
             try
             {
                 await _service.Update(request);
             }
+            catch (EntityNotFoundException<HotelType> ex)
+            {
+                return NotFound(new ErrorResponse(ex.Message));
+            }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error while processing request from {Hotel Types}", request);
+                return BadRequest(new ErrorResponse(ex.Message));
             }
 
-            return new Response(200, true, null);
+            return Ok();
         }
 
         [HttpDelete]
         [Route("deleteById")]
-        public async Task<Response> Delete(int id)
+        public async Task<ActionResult> Delete(int id)
         {
+            HotelType hotelType = await _service.GetByIdAsync(id);
+            if (hotelType == null)
+                return NotFound(new ErrorResponse("Hotel Type not found"));
             await _service.Delete(id);
 
-            return new Response(200, true, null);
+            return NoContent();
         }
 
         [HttpGet]
         [Route("GetById")]
         [Authorize]
-        public async Task<HotelTypeResponse> GetById(int id)
+        public async Task<ActionResult> GetById(int id)
         {
             HotelType hotelType = await _service.GetByIdAsync(id);
-
-            return new HotelTypeResponse(200, true, null, hotelType);
+            if (hotelType == null)
+                return NotFound(new ErrorResponse("Hotel Type not found"));
+            return Ok(new HotelTypeDto(hotelType));
         }
 
         [HttpGet]
         [Route("GetAll")]
         [Authorize]
-        public async Task<HotelTypesListResponse> GetAll()
+        public async Task<ActionResult<List<HotelTypeDto>>> GetAll()
         {
             var list = await _service.GetAll();
 
-            return new HotelTypesListResponse(200, true, null, list);
+            return Ok(list.Select(dep => new HotelTypeDto(dep)).ToList());
+
         }
     }
 }
